Revert "Merge branch 'master' of https://git.mif.vu.lt/mamu8341/gis"
This reverts commit fe64dd1899de8a5231a217b5f49346869ec723a1

+++ src/app/src/main/java/com/example/leisuremap/Object.java
@@ -5,13 +5,13 @@ public class Object {
     private double distance;
     private double lat;
     private double lon;
-    private double rating;
+    private int rating;
     private String city;
     private String type;
     private double score;
     private String distanceString;
 
-    public Object(String name, double distance, double lat, double lon, double rating, String city, String type, double score) {
+    public Object(String name, double distance, double lat, double lon, int rating, String city, String type, double score) {
         this.name = name;
         this.distance = distance;
         this.lat = lat;
@@ -70,11 +70,11 @@ public class Object {
         this.score = score;
     }
 
-    public double getRating() {
+    public int getRating() {
         return rating;
     }
 
-    public void setRating(double rating) {
+    public void setRating(int rating) {
         this.rating = rating;
     }
